Move enums and typedefs into tyrant.{cpp,h}
